Consolidate focus and calib views
RE #39352
