+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/Engineering/gui/engineering_diffraction/tabs/calibration/view.py
@@ -31,6 +31,10 @@ class CalibrationView(QtWidgets.QWidget, Ui_calib):
         self.finder_path.setEnabled(False)
         self.finder_path.setFileExtensions([".prm"])
 
+        self.finder_focus.setLabelText("Sample Run #")
+        self.finder_focus.setInstrumentOverride(instrument)
+        self.finder_focus.allowMultipleFiles(True)
+
     # =================
     # Slot Connectors
     # =================
@@ -72,6 +76,7 @@ class CalibrationView(QtWidgets.QWidget, Ui_calib):
     def set_instrument_override(self, instrument):
         self.finder_sample.setInstrumentOverride(instrument)
         self.finder_vanadium.setInstrumentOverride(instrument)
+        self.finder_focus.setInstrumentOverride(instrument)
 
     def set_sample_enabled(self, set_to):
         self.finder_sample.setEnabled(set_to)
@@ -151,7 +156,7 @@ class CalibrationView(QtWidgets.QWidget, Ui_calib):
     # =================
 
     def is_searching(self):
-        return self.finder_sample.isSearching() or self.finder_vanadium.isSearching()
+        return self.finder_sample.isSearching() or self.finder_vanadium.isSearching() or self.finder_focus.isSearching()
 
     # =================
     # Force Actions
@@ -177,3 +182,50 @@ class CalibrationView(QtWidgets.QWidget, Ui_calib):
         self.setTabOrder(self.check_roiCalib, self.widget_cropping)
         self.setTabOrder(self.widget_cropping, self.check_plotOutput)
         self.setTabOrder(self.check_plotOutput, self.button_calibrate)
+
+    # =================
+    # Focus Slot Connectors
+    # =================
+
+    def set_on_focus_clicked(self, slot):
+        self.button_focus.clicked.connect(slot)
+
+    # =================
+    # Focus Component Setters
+    # =================
+
+    def set_focus_button_enabled(self, enabled):
+        self.button_focus.setEnabled(enabled)
+
+    def set_plot_output_enabled(self, enabled):
+        self.check_plotOutput.setEnabled(enabled)
+
+    def set_region_display_text(self, text):
+        self.regionDisplay.setText(text)
+
+    # =================
+    # Focus Component Getters
+    # =================
+
+    def get_focus_filenames(self):
+        return self.finder_focus.getFilenames()
+
+    def get_focus_valid(self):
+        return self.finder_focus.isValid()
+
+    def get_focus_plot_output(self):
+        return self.check_focusPlotOutput.isChecked()
+
+    # =================
+    # Focus Internal Setup
+    # =================
+
+    def set_default_files(self, filepaths, directory):
+        if not filepaths:
+            return
+        self.finder_focus.setUserInput(",".join(filepaths))
+        if directory:
+            self.set_finder_last_directory(directory)
+
+    def set_finder_last_directory(self, directory):
+        self.finder_focus.setLastDirectory(directory)
